+++ leptos-fluent/src/lib.rs
@@ -251,6 +251,8 @@ pub mod localstorage;
 #[doc(hidden)]
 pub mod url;
 
+#[cfg(feature = "system")]
+pub use current_locale::current_locale;
 #[doc(hidden)]
 pub use web_sys;